Split out GHC-independent utilities into separate module (#2)
At some point we're going to have to support more versions of GHC
other than 9.6 and the logic for doing that could potentially be
subtly different in each case.  However, we want to minimize the amount
of code that we have to copy and paste and tweak to support each
version of GHC.

To that end, this change splits out the code bits that don't depend on
the GHC API so that they can be reused by the plugin for each supported
version of GHC.

After this change all the open telemetry logic is handled in the
`OpenTelemetry.Plugin.Shared` module and all the GHC logic is handled
in the `OpenTelemetry.Plugin` module.

I also took this opportunity to improve the completeness and quality of
the haddocks.