Add support for sampling a subset of module spans (#3)
In our internal codebase we have 6000+ modules which each generate a span and
we also typically have two compiler phases per module which each generate two
sub-spans, so we generate roughly 3 × 6000+ = 18000+ spans that we export as
part of each trace if we don't do any sampling.

Unfortunately, Honeycomb cannot render a trace with 18000 spans, so we need
some way to sample a random subset of them so that we can still get some useful
insights about our build.

Fortunately, open telemetry does support sampling things, but none of the stock
samplers work for our use case.  In particular, the only relevant sampler that
open telemetry provides out of the box is a `traceidratio` sampler which
samples a subset of traces (but not spans), so that wouldn't really fix our
problem.

So what I did was have our plugin interpret an optional `spanratio` sampler
which can be configured through the standard `OTEL_TRACES_SAMPLER{,_ARG}`
environment variables:

https://opentelemetry.io/docs/specs/otel/configuration/sdk-environment-variables/#general-sdk-configuration

So now if you export something like:

```bash
OTEL_TRACES_SAMPLER=spanratio
OTEL_TRACES_SAMPLER_ARG=0.1
```

… then the plugin will sample 10% of all modules.

Note that we don't apply the span ratio to sub-spans of a module (which would
generate weird traces) and we don't apply the span ratio to the root span
that envelopes the entire trace (because then you'd drop a whole lot of
traces).  We only apply the span ratio to the module-wide spans.

Note that this required upgrading the `hs-opentelemetry-api` dependency because
the latest version changed the API of the `attributes` field of
`SpanArguments` and I'm only really interested in supporting the latest
version of the API and not older versions.  That's why the
`hs-opentelemetry-api` now has a lower bound, too.