Support GHC 9.2 and later (#4)
It turns out that with a few small adjustments the GHC-specific code
can be made to work with all GHC versions since 9.2.  It's not
possible to support GHC 9.0 or older, at all, because the
`Plugin` type is missing the `driverPlugin` field which we depend on
to initialize things.

I still kept the GHC-specific code in a separate module so that if
we need to go back to supporting one GHC module per GHC version then
it'll still be easy to do that.