Make initialization idempotent
When testing this plugin internally I ran into a hang in the context of
`haskell-language-server` where it would hang due to running the
plugin initialization logic more than once.  This leads to the same
`putMVar` command being called twice and the second invocation hangs
(since the `MVar` is now full).

This changes all of the `putMVar`s to be `tryPutMVar`s so that they're
now idempotent and avoid this issue.